+++ docs/make.jl
@@ -11,7 +11,8 @@ makedocs(
         prettyurls = DEPLOYDOCS,
     ),
     authors = "Éric Thiébaut and contributors",
-    pages = ["index.md", "install.md", "starting.md", "requests.md", "connect.md", "library.md"]
+    pages = ["index.md", "install.md", "starting.md", "requests.md",
+             "connect.md", "drawing.md", "library.md"]
 )
 
 if DEPLOYDOCS